+++ src/main.rs
@@ -1,9 +1,10 @@
 #![warn(unused_extern_crates)]
+#![allow(special_module_name)]
 use std::path::{Path, PathBuf};
 
 use crate::lib::AnyhowResult;
 use anyhow::Context;
-use bip39::Mnemonic;
+use bip39::{Language, Mnemonic};
 use clap::{crate_version, Args, Parser};
 use lib::AuthInfo;
 
@@ -14,6 +15,8 @@ mod lib;
 #[derive(Parser)]
 #[clap(name("quill"), version = crate_version!())]
 pub struct CliOpts {
+    #[clap(flatten, next_help_heading = "COMMON")]
+    global_opts: GlobalOpts,
     #[clap(subcommand)]
     command: commands::Command,
 }
@@ -21,58 +24,46 @@ pub struct CliOpts {
 #[derive(Args)]
 struct GlobalOpts {
     /// Path to your PEM file (use "-" for STDIN)
-    #[clap(long)]
+    #[clap(long, group = "auth", global = true)]
     pem_file: Option<PathBuf>,
 
-    #[clap(long)]
+    #[clap(long, group = "auth", global = true)]
     hsm: bool,
 
-    #[clap(long)]
+    #[clap(long, global = true)]
     hsm_libpath: Option<PathBuf>,
 
-    #[clap(long)]
+    #[clap(long, global = true)]
     hsm_slot: Option<usize>,
 
-    #[clap(long)]
+    #[clap(long, global = true)]
     hsm_id: Option<String>,
 
     /// Path to your seed file (use "-" for STDIN)
-    #[clap(long)]
+    #[clap(long, global = true)]
     seed_file: Option<PathBuf>,
 
     /// Output the result(s) as UTF-8 QR codes.
-    #[clap(long)]
+    #[clap(long, global = true)]
     qr: bool,
 
     /// Fetches the root key before making requests so that interfacing with local instances is possible.
     /// DO NOT USE WITH ANY REAL INFORMATION
-    #[clap(long = "insecure-local-dev-mode", name = "insecure-local-dev-mode")]
+    #[clap(
+        long = "insecure-local-dev-mode",
+        name = "insecure-local-dev-mode",
+        global = true
+    )]
     fetch_root_key: bool,
 }
 
-#[derive(Args)]
-pub struct BaseOpts<T: Args> {
-    #[clap(flatten)]
-    command_opts: T,
-    #[clap(flatten, next_help_heading = "COMMON")]
-    global_opts: GlobalOpts,
-}
-
-fn main() {
+fn main() -> AnyhowResult {
     let opts = CliOpts::parse();
-    if let Err(err) = commands::dispatch(opts.command) {
-        for (level, cause) in err.chain().enumerate() {
-            if level == 0 {
-                eprintln!("Error: {}", err);
-                continue;
-            }
-            if level == 1 {
-                eprintln!("Caused by:");
-            }
-            eprintln!("{:width$}{}", "", cause, width = level * 2);
-        }
-        std::process::exit(1);
-    }
+    let qr = opts.global_opts.qr;
+    let fetch_root_key = opts.global_opts.fetch_root_key;
+    let auth = get_auth(opts.global_opts)?;
+    commands::dispatch(&auth, opts.command, fetch_root_key, qr)?;
+    Ok(())
 }
 
 fn get_auth(opts: GlobalOpts) -> AnyhowResult<AuthInfo> {
@@ -114,7 +105,8 @@ fn read_pem(pem_file: Option<&Path>, seed_file: Option<&Path>) -> AnyhowResult<O
 }
 
 fn parse_mnemonic(phrase: &str) -> AnyhowResult<Mnemonic> {
-    Mnemonic::parse(phrase).context("Couldn't parse the seed phrase as a valid mnemonic. {:?}")
+    Mnemonic::from_phrase(phrase, Language::English)
+        .context("Couldn't parse the seed phrase as a valid mnemonic. {:?}")
 }
 
 fn read_file(path: impl AsRef<Path>, name: &str) -> AnyhowResult<String> {
@@ -135,7 +127,7 @@ fn read_file(path: impl AsRef<Path>, name: &str) -> AnyhowResult<String> {
 #[cfg(test)]
 mod tests {
     use crate::read_pem;
-    use bip39::Mnemonic;
+    use bip39::{Language, Mnemonic};
 
     #[test]
     fn test_read_pem_none_none() {
@@ -169,7 +161,9 @@ mod tests {
         seed_file
             .write_all(phrase.as_bytes())
             .expect("Cannot write to temp file");
-        let mnemonic = crate::lib::mnemonic_to_pem(&Mnemonic::parse(phrase).unwrap()).unwrap();
+        let mnemonic =
+            crate::lib::mnemonic_to_pem(&Mnemonic::from_phrase(phrase, Language::English).unwrap())
+                .unwrap();
 
         let pem = read_pem(None, Some(seed_file.path()))
             .expect("Unable to read seed_file")