+++ tests/commands/sns-list-deployed-snses.sh
@@ -1,0 +1,1 @@
+"$QUILL" sns list-deployed-snses --canister-ids-file ./sns_canister_ids.json --dry-run