+++ tests/commands/sns-refund.sh
@@ -1,0 +1,1 @@
+"$QUILL" sns get-swap-refund --canister-ids-file ./sns_canister_ids.json --pem-file - | "$QUILL" send --dry-run -