+++ tests/commands/sns-status.sh
@@ -1,0 +1,1 @@
+"$QUILL" sns status --dry-run --canister-ids-file ./sns_canister_ids.json